Feat : order payment 타입 분리
