+++ src/types/order.ts
@@ -2,8 +2,6 @@ import { CartItemDto } from '@/types/product'
 
 export type OrderItem = Pick<CartItemDto, 'itemId' | 'itemName' | 'itemPrice' | 'itemCount' | 'discountRate'>
 
-export type PaymentMethodType = 'KAKAOPAY'
-
 export interface IOrder {
   orderId: number
   amount: number
@@ -13,13 +11,6 @@ export interface IOrder {
   phoneNumber: string | null
   address: null
   orderItemDtos: OrderItem[]
-}
-
-export interface IPaymentParams {
-  // memberId: number
-  orderId: number
-  totalPrice: number
-  paymentMethod: PaymentMethodType
 }
 
 export const ShippingStatus = {
